cleanup and added comments

+++ src/main/java/com/madhu/core/PersonMapper.java
@@ -13,7 +13,7 @@ public class PersonMapper implements ResultSetMapper<Person> {
 
     public Person map(int index, ResultSet r, StatementContext ctx) throws SQLException {
         Person person = new Person();
-        person.setId(r.getLong("ID"));
+        person.setId(r.getInt("ID"));
         person.setName(r.getString("NAME"));
         return person;
     }
